Two digits combin

+++ 0x01-variables_if_else_while/100-print_comb3.c
@@ -5,7 +5,7 @@
  **/
 int main(void)
 {
-	unsigned short int x, y, z;
+	int x, y, z;
 
 	z = x + y;
